+++ pkg/keyvisual/core.go
@@ -1,0 +1,114 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package keyvisual
+
+import (
+	"encoding/hex"
+	"net/http"
+	"strconv"
+	"time"
+
+	"github.com/gin-gonic/gin"
+	"github.com/pingcap/log"
+	"go.uber.org/zap"
+
+	// Import for swag go doc
+	_ "github.com/pingcap-incubator/tidb-dashboard/pkg/apiserver/utils"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/matrix"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/region"
+	"github.com/pingcap-incubator/tidb-dashboard/pkg/keyvisual/storage"
+)
+
+type ServiceCore struct {
+	maxDisplayY int
+
+	stat     *storage.Stat
+	strategy matrix.Strategy
+}
+
+// @Summary Key Visual Heatmaps
+// @Description Heatmaps in a given range to visualize TiKV usage
+// @Produce json
+// @Param startkey query string false "The start of the key range"
+// @Param endkey query string false "The end of the key range"
+// @Param starttime query int false "The start of the time range (Unix)"
+// @Param endtime query int false "The end of the time range (Unix)"
+// @Param type query string false "Main types of data" Enums(written_bytes, read_bytes, written_keys, read_keys, integration)
+// @Success 200 {object} matrix.Matrix
+// @Router /keyvisual/heatmaps [get]
+// @Security JwtAuth
+// @Failure 401 {object} utils.APIError "Unauthorized failure"
+func (s *ServiceCore) heatmapsHandler(c *gin.Context) {
+	startKey := c.Query("startkey")
+	endKey := c.Query("endkey")
+	startTimeString := c.Query("starttime")
+	endTimeString := c.Query("endtime")
+	typ := c.Query("type")
+
+	endTime := time.Now()
+	startTime := endTime.Add(-360 * time.Minute)
+	if startTimeString != "" {
+		tsSec, err := strconv.ParseInt(startTimeString, 10, 64)
+		if err != nil {
+			log.Error("parse ts failed", zap.Error(err))
+			c.JSON(http.StatusBadRequest, "bad request")
+			return
+		}
+		startTime = time.Unix(tsSec, 0)
+	}
+	if endTimeString != "" {
+		tsSec, err := strconv.ParseInt(endTimeString, 10, 64)
+		if err != nil {
+			log.Error("parse ts failed", zap.Error(err))
+			c.JSON(http.StatusBadRequest, "bad request")
+			return
+		}
+		endTime = time.Unix(tsSec, 0)
+	}
+	if !(startTime.Before(endTime) && (endKey == "" || startKey < endKey)) {
+		c.JSON(http.StatusBadRequest, "bad request")
+		return
+	}
+
+	log.Debug("Request matrix",
+		zap.Time("start-time", startTime),
+		zap.Time("end-time", endTime),
+		zap.String("start-key", startKey),
+		zap.String("end-key", endKey),
+		zap.String("type", typ),
+	)
+
+	if startKeyBytes, err := hex.DecodeString(startKey); err == nil {
+		startKey = string(startKeyBytes)
+	} else {
+		c.JSON(http.StatusBadRequest, "bad request")
+		return
+	}
+	if endKeyBytes, err := hex.DecodeString(endKey); err == nil {
+		endKey = string(endKeyBytes)
+	} else {
+		c.JSON(http.StatusBadRequest, "bad request")
+		return
+	}
+	baseTag := region.IntoTag(typ)
+	plane := s.stat.Range(startTime, endTime, startKey, endKey, baseTag)
+	resp := plane.Pixel(s.strategy, s.maxDisplayY, region.GetDisplayTags(baseTag))
+	resp.Range(startKey, endKey)
+	// TODO: An expedient to reduce data transmission, which needs to be deleted later.
+	resp.DataMap = map[string][][]uint64{
+		typ: resp.DataMap[typ],
+	}
+	// ----------
+	c.JSON(http.StatusOK, resp)
+}
